\name{getsimexest}
\alias{getsimexest}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getsimexest 
}
\description{
extract the estimates for every lambda}
\usage{
getsimexest(indata)
}

\arguments{
  \item{indata}{swgee object from the function swgee}
}

\details{
internal function for the extrapolation step
}

\references{
Cook, J.R. and Stefanski, L.A. (1994) Simulation-extrapolation estimation in parametric measurement error models. \emph{Journal of the American Statistical Association}, 89, 1314-1328.

Carrol, R.J., Ruppert, D., Stefanski, L.A. and Crainiceanu, C. (2006) Measurement error in nonlinear models: A modern perspective., Second Edition. London: Chapman and Hall.

Yi, G. Y. (2008) A simulation-based marginal method for longitudinal data with dropout and mismeasured covariates. \emph{Biostatistics}, 9, 501-512.
}
 
\author{
Juan Xiong<jxiong@szu.edu.cn>, Grace Y. Yi<yyi@uwaterloo.ca>
}



