\name{print.summary.swgee}
\alias{print.summary.swgee}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
print.summary.swgee
}
\description{
Summary method for class "swgee" 
}
\usage{
## S3 method for class 'swgee'
\method{print}{summary.swgee}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
object of class 'swgee'
}
  \item{\dots}{
further arguments passed to or from other functions.
}
}

\value{
Print summary nicely
}

\references{
Cook, J.R. and Stefanski, L.A. (1994) Simulation-extrapolation estimation in parametric measurement error models. \emph{Journal of the American Statistical Association}, 89, 1314-1328.

Carrol, R.J., Ruppert, D., Stefanski, L.A. and Crainiceanu, C. (2006) Measurement error in nonlinear models: A modern perspective., Second Edition. London: Chapman and Hall.

Yi, G. Y. (2008) A simulation-based marginal method for longitudinal data with dropout and mismeasured covariates. \emph{Biostatistics}, 9, 501-512.
}

\author{
Juan Xiong<jxiong@szu.edu.cn>, Grace Y. Yi<yyi@uwaterloo.ca>
}

