\name{summary.swgee}
\alias{summary.swgee}
\title{
summary.swgee
}
\description{
Summary method for class "swgee"
}
\usage{
\method{summary}{swgee}(object, ...)
}
\arguments{
  \item{object}{
object of class 'swgee'
}
  \item{\dots}{
further arguments passed to or from other functions.
}
}
\details{
The function summary.swgee computes and returns a list of summary statistics of the response process and missing process
}
\value{
summary estimates for parameters associated with response process and missing process
}

\references{
Cook, J.R. and Stefanski, L.A. (1994) Simulation-extrapolation estimation in parametric measurement error models. \emph{Journal of the American Statistical Association}, 89, 1314-1328.

Carrol, R.J., Ruppert, D., Stefanski, L.A. and Crainiceanu, C. (2006) Measurement error in nonlinear models: A modern perspective., Second Edition. London: Chapman and Hall.

Yi, G. Y. (2008) A simulation-based marginal method for longitudinal data with dropout and mismeasured covariates. \emph{Biostatistics}, 9, 501-512.
}

\author{
Juan Xiong<jxiong@szu.edu.cn>, Grace Y. Yi<yyi@uwaterloo.ca>
}


