\name{swgee-package}
\alias{swgee-package}
\docType{package}
\title{
\packageTitle{swgee}
}
\description{
\packageDescription{swgee}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{swgee}
\packageIndices{swgee}
Implementation of the SIMEX inverse probability weighted GEE method for longitudinal data with missing observations and measurement error in covariates
}

\author{
Juan Xiong <jxiong@szu.edu.cn>, Grace Y. Yi<yyi@uwaterloo.ca>

Maintainer: Juan Xiong <jxiong@szu.edu.cn>
}

\references{
Cook, J.R. and Stefanski, L.A. (1994) Simulation-extrapolation estimation in parametric measurement error models. \emph{Journal of the American Statistical Association}, 89, 1314-1328.

Carrol, R.J., Ruppert, D., Stefanski, L.A. and Crainiceanu, C. (2006) Measurement error in nonlinear models: A modern perspective., Second Edition. London: Chapman and Hall.

Yi, G. Y. (2008) A simulation-based marginal method for longitudinal data with dropout and mismeasured covariates. \emph{Biostatistics}, 9, 501-512.
}
\seealso{\code{\link[geepack]{geeglm}}}

